add accuracy fix

+++ rpm-sensor/src/main/java/RPMrmiInterface/RPMInterface.java
@@ -10,7 +10,7 @@ public interface RPMInterface extends Remote {
 
     public String ping() throws RemoteException;
 
-    public static int getRPMValue() throws RemoteException {
+    public static double getRPMValue() throws RemoteException {
         return 0;
     }
 }
